Flight Price Prediction Project
This project is about predicting the price of flight inside India. The dataset has been taken from Kaggle which is based on different cities of India. The model is trained on RandomForestRegressor and its accuracy is about 95 percent. Further this model has been integrated with Flask backend framework of Python. For user interaction simple webpage has been created using HTML, CSS and JS